delete visit

+++ src/app/patient-visits/patient-visits.component.ts
@@ -3,7 +3,7 @@ import { PatientService } from '../services/patientByPhoneNumber.service';
 import { GetPatientVisitDto } from '../types/GetPatientVisitDto';
 import { PatientService2 } from '../services/patient.service';
 import { empty } from 'rxjs';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-patient-visits',
@@ -31,20 +31,39 @@ patientVisits?: GetPatientVisitDto
     });
   }
   delete(e: Event , id: number){
-    this.patientService.deleteAppointment(id).subscribe({
-      next:()=>{
-        this.patientService.GetPatientVisitsByPhone(this.patientPhoneNumbaer!).subscribe({
-          next: (patientVisits) => {
-            this.patientVisits = patientVisits;
-          },
-          error: (error) => {
-            console.error('Error fetching visits:', error);
-          },
-        });
-      },
-      error:(error)=>{
-        console.log("delete patient visit api failed",error)
-      }
-    })
+    this.showConfirmation(id);
   }
+
+  showConfirmation( id: number) {
+    Swal.fire({
+      title: 'Are you sure you want to delete vist?',
+      text: 'You won\'t be able to revert this!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+    }).then((result) => {
+      if (result.value) {
+        this.patientService.deleteAppointment(id).subscribe({
+          next:()=>{
+            this.patientService.GetPatientVisitsByPhone(this.patientPhoneNumbaer!).subscribe({
+              next: (patientVisits) => {
+                this.patientVisits = patientVisits;
+              },
+              error: (error) => {
+                console.error('Error fetching visits:', error);
+              },
+            });
+          },
+          error:(error)=>{
+            console.log("delete patient visit api failed",error)
+          }
+        })
+        Swal.fire('Deleted!', 'Your visit has been deleted.', 'success');
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire('Cancelled', 'Your file is safe :)', 'info');
+      }
+    });
+  }
+  
 }
